refactor(extension): decompose oversized files + share chip-overflow hook
Pure, behavior-preserving cleanup of the branch's structural debt.

- design-system: extract a headless chip-overflow into
  packages/design-system (pure computeChipsShown + useChipOverflow +
  useAnchoredTooltip) and consume it from explorer FeedCardView and
  extension ContextPills, killing the duplicated single-line overflow
  algorithm.
- WeightModal 1208->669: split the Amplify/basket/reward surfaces into
  weight/WeightAmplifyTicket, WeightBasketRow and WeightRewardTicket;
  shared helpers moved to weight/types.ts.
- CircleFeedTab 1194->724: extract the trust-circle/cert/feed-item/context
  data pipeline into a useCircleFeed hook; shared types into feedTypes.ts.
- GroupDetailView 852->537: extract UrlRow, GroupDetailHeader and
  LevelProgress into a group-detail/ folder.
- tests: +36 (chipOverflow, circleFeedSort, reputationBackersService,
  circleBuilders); explorer suite 97->133.

No extension component exceeds 800 lines now. Verified: explorer tsc -b +
vitest 133/133, design-system vitest 17/17, extension tsc 0 errors.
