+++ apps/extension/components/modals/WeightModal.tsx
@@ -1,5 +1,3 @@
-import { UserBadge, VerbTag } from "@0xsofia/design-system"
-import type { IntentionSlug, UserBadgeTier } from "@0xsofia/design-system"
 import { useEffect, useMemo, useState } from "react"
 import { createPortal } from "react-dom"
 import { formatUnits, getAddress } from "viem"
@@ -15,54 +13,27 @@ import {
   useWalletFromStorage
 } from "~/hooks"
 import type { ModalTriplet } from "~/hooks"
-import { EXPLORER_URLS } from "~/lib/config/chainConfig"
-import { contextLabel } from "~/lib/config/contextDisplay"
-import ContextPills from "../ui/ContextPills"
-import { createHookLogger, getFaviconUrl } from "~/lib/utils"
-import {
-  INTENTION_CONFIG,
-  predicateLabelToIntentionType
-} from "~/types/intentionCategories"
-import type { IntentionType } from "~/types/intentionCategories"
+import { createHookLogger } from "~/lib/utils"
 
-import contributorBadge from "../ui/img/badges/contributor.png"
-import explorerBadge from "../ui/img/badges/explorer.png"
-import pioneerBadge from "../ui/img/badges/pioneer.png"
 import SofiaLoader from "../ui/SofiaLoader"
 import {
   DEFAULT_WEIGHT,
   formatTrust,
   getWeightValue,
-  weightOptions,
   type DiscoveryReward,
   type WeightOptionId
 } from "./weight/types"
 import WeightActions from "./weight/WeightActions"
+import WeightAmplifyTicket from "./weight/WeightAmplifyTicket"
 import WeightApplyAllBar from "./weight/WeightApplyAllBar"
 import WeightItem from "./weight/WeightItem"
+import WeightRewardTicket from "./weight/WeightRewardTicket"
 import WeightSuccessCard from "./weight/WeightSuccessCard"
 import WeightTotal from "./weight/WeightTotal"
 
 import "../styles/Modal.css"
 
 const logger = createHookLogger("WeightModal")
-
-const TIER_ORDER = ["pioneer", "explorer", "contributor"] as const
-type RewardTier = (typeof TIER_ORDER)[number]
-
-const BADGE_IMAGES: Record<UserBadgeTier, string> = {
-  pioneer: pioneerBadge,
-  explorer: explorerBadge,
-  contributor: contributorBadge
-}
-
-const hostFromUrl = (url: string): string => {
-  try {
-    return new URL(url).hostname.replace(/^www\./, "")
-  } catch {
-    return url
-  }
-}
 
 interface WeightModalProps {
   isOpen: boolean
@@ -651,553 +622,43 @@ const WeightModal = ({
           )}
 
         {showAmpForm && (
-          <div className="amp b3 b3--v5">
-            <div className="b3-bg b3-bg--v5" aria-hidden="true">
-              <div className="b3-bg-hex b3-bg-hex--1" />
-              <div className="b3-bg-hex b3-bg-hex--2" />
-              <div className="b3-bg-hex b3-bg-hex--3" />
-              <div className="b3-bg-grain" />
-            </div>
-
-            {isProcessing && (
-              <div className="b3-loading" role="status" aria-live="polite">
-                <SofiaLoader size={120} />
-                <p className="b3-loading-title">Amplifying</p>
-                <p className="b3-loading-step">{processingStep}</p>
-                <p className="b3-loading-warn">
-                  Do not close or navigate away from this tab
-                </p>
-              </div>
-            )}
-
-            <div className="b3-ticket">
-              <div className="b3-body">
-                <div className="b3-headline">
-                  <span className="b3-drop">A</span>
-                  <h1 className="b3-h1">mplify.</h1>
-                </div>
-                <p className="b3-sub">
-                  {fixedDeposit != null
-                    ? "Review the cost breakdown before confirming."
-                    : "Compose your basket — pick a strength for each mark."}
-                </p>
-
-                {fixedDeposit == null && activeCount > 1 && (
-                  <div className="b3-presets">
-                    <span className="b3-presets-k">QUICK</span>
-                    {weightOptions.map((o) => (
-                      <button
-                        key={o.id}
-                        className="b3-preset"
-                        type="button"
-                        disabled={isProcessing}
-                        onClick={() => handleApplyAll(o.id)}>
-                        All · {o.label}
-                      </button>
-                    ))}
-                  </div>
-                )}
-
-                <div className="b3-basket">
-                  {triplets.map((triplet, index) => {
-                    const on = !removedIndices.has(index)
-                    const sel = selectedWeights[index]
-                    const name = triplet.description || triplet.triplet.object
-                    const intentKey: IntentionType | null = triplet.intention
-                      ? (((triplet.intention as string) in INTENTION_CONFIG
-                          ? triplet.intention
-                          : triplet.intention.replace(
-                              /^for_/,
-                              ""
-                            )) as IntentionType)
-                      : // Trust / distrust rows carry no `intention` —
-                        // resolve the verb from the predicate label.
-                        predicateLabelToIntentionType(triplet.triplet.predicate)
-                    const intentEntry =
-                      intentKey && intentKey in INTENTION_CONFIG
-                        ? INTENTION_CONFIG[intentKey]
-                        : null
-                    const ctxSlugs =
-                      triplet.interestContexts &&
-                      triplet.interestContexts.length
-                        ? triplet.interestContexts
-                        : triplet.interestContext
-                          ? [triplet.interestContext]
-                          : []
-                    const canToggleOff = activeCount > 1
-                    const lockedSingle = triplets.length <= 1
-                    const platform =
-                      ppEnabled && fixedDeposit == null
-                        ? detectPlatformFromUrl(triplet.url)
-                        : null
-                    const toggle = () => {
-                      if (lockedSingle || isProcessing) return
-                      if (on) {
-                        if (!canToggleOff) return
-                        setRemovedIndices((prev) => new Set(prev).add(index))
-                        onRemoveTriplet?.(triplet.id)
-                      } else {
-                        setRemovedIndices((prev) => {
-                          const next = new Set(prev)
-                          next.delete(index)
-                          return next
-                        })
-                      }
-                    }
-                    const rowTrust = on ? getWeightValue(sel) : 0
-                    const ppFraction = platform
-                      ? getPpForSlug(platform.slug) / PP_FEE_DENOMINATOR
-                      : 0
-                    const platAmt = on ? rowTrust * ppFraction : 0
-                    const urlAmt = on ? rowTrust - platAmt : 0
-                    return (
-                      <div
-                        className={`b3-row${on ? " is-on" : ""}`}
-                        key={triplet.id}>
-                        <button
-                          className="b3-check"
-                          role="checkbox"
-                          aria-checked={on}
-                          aria-label={on ? "Deselect mark" : "Select mark"}
-                          type="button"
-                          disabled={lockedSingle || (on && !canToggleOff)}
-                          onClick={toggle}>
-                          {on && (
-                            <svg
-                              width="11"
-                              height="11"
-                              viewBox="0 0 24 24"
-                              fill="none">
-                              <path
-                                d="M5 12l4 4L19 6"
-                                stroke="var(--ds-on-accent)"
-                                strokeWidth="3"
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                              />
-                            </svg>
-                          )}
-                        </button>
-
-                        <img
-                          src={getFaviconUrl(triplet.url, 32)}
-                          alt=""
-                          className="b3-row-fav"
-                          onError={(e) => {
-                            ;(e.target as HTMLImageElement).style.visibility =
-                              "hidden"
-                          }}
-                        />
-
-                        <div className="b3-row-meta">
-                          <div className="b3-row-name" title={name}>
-                            {name}
-                          </div>
-                          <div className="b3-row-sub" title={triplet.url}>
-                            {hostFromUrl(triplet.url)}
-                          </div>
-                        </div>
-
-                        <div className="b3-row-trust">
-                          <span className="b3-row-trust-val">
-                            {fixedDeposit != null
-                              ? "—"
-                              : formatTrust(rowTrust)}
-                          </span>
-                          <span className="b3-row-trust-unit">TRUST</span>
-                        </div>
-
-                        {/* Verb + context pills get their own full-width grid
-                            row (grid-column 1/-1) so multiple contexts wrap
-                            cleanly instead of crowding the title column. */}
-                        {(intentEntry || ctxSlugs.length > 0) && (
-                          <div className="b3-row-tags">
-                            {intentEntry && (
-                              <VerbTag
-                                intent={intentKey as IntentionSlug}
-                                label={intentEntry.label}
-                              />
-                            )}
-                            <ContextPills slugs={ctxSlugs} />
-                          </div>
-                        )}
-
-                        {/* Weight selector spans its own full-width row below
-                            the tags so the 3 tiers stay uncramped. */}
-                        {fixedDeposit == null && (
-                          <div
-                            className="b3-tiers"
-                            role="radiogroup"
-                            aria-disabled={!on}>
-                            {weightOptions.map((o) => (
-                              <button
-                                key={o.id}
-                                role="radio"
-                                aria-checked={sel === o.id}
-                                type="button"
-                                className={`b3-tier${
-                                  sel === o.id ? " is-on" : ""
-                                }`}
-                                disabled={!on || isProcessing}
-                                onClick={() =>
-                                  on && handleWeightSelection(index, o.id)
-                                }>
-                                <span className="b3-tier-label">{o.label}</span>
-                                <span className="b3-tier-value">{o.value}</span>
-                              </button>
-                            ))}
-                          </div>
-                        )}
-
-                        {platform && fixedDeposit == null && (
-                          <div className={`b3-dest${on ? "" : " is-off"}`}>
-                            <div className="b3-dest-line">
-                              <span className="b3-dest-tag b3-dest-tag--url">
-                                URL
-                              </span>
-                              <span
-                                className="b3-dest-where"
-                                title={triplet.url}>
-                                {hostFromUrl(triplet.url)} · this page
-                              </span>
-                              <span className="b3-dest-amt">
-                                {formatTrust(urlAmt)} T
-                              </span>
-                            </div>
-                            <div className="b3-dest-line">
-                              <span className="b3-dest-tag b3-dest-tag--domain">
-                                DOMAIN
-                              </span>
-                              <span className="b3-dest-where">
-                                {platform.label} · whole platform
-                              </span>
-                              <span className="b3-dest-ctl">
-                                <input
-                                  type="range"
-                                  min={0}
-                                  max={100000}
-                                  step={1000}
-                                  value={getPpForSlug(platform.slug)}
-                                  onChange={(e) =>
-                                    setPpForSlug(
-                                      platform.slug,
-                                      Number(e.target.value)
-                                    )
-                                  }
-                                  disabled={!on || isProcessing}
-                                  aria-label={`${platform.label} share`}
-                                  className="b3-dest-range"
-                                  style={
-                                    {
-                                      "--pp-fill": `${
-                                        getPpForSlug(platform.slug) / 1000
-                                      }%`
-                                    } as React.CSSProperties
-                                  }
-                                />
-                                <span className="b3-dest-pct">
-                                  {(getPpForSlug(platform.slug) / 1000).toFixed(
-                                    0
-                                  )}
-                                  %
-                                </span>
-                              </span>
-                              <span className="b3-dest-amt">
-                                {formatTrust(platAmt)} T
-                              </span>
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    )
-                  })}
-                </div>
-
-                {curveSelector && (
-                  <div className="b3-extra">
-                    <div className="curve-selector">
-                      <span className="curve-selector-label">
-                        Bonding curve:
-                      </span>
-                      <div className="curve-toggle">
-                        <button
-                          className={`curve-toggle-btn ${
-                            curveSelector.selected === "linear" ? "active" : ""
-                          }`}
-                          onClick={() => curveSelector.onChange("linear")}
-                          disabled={isProcessing}>
-                          Linear
-                        </button>
-                        <button
-                          className={`curve-toggle-btn ${
-                            curveSelector.selected === "progressive"
-                              ? "active"
-                              : ""
-                          }`}
-                          onClick={() => curveSelector.onChange("progressive")}
-                          disabled={isProcessing}>
-                          Progressive
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-                )}
-
-                {gsEnabled && (
-                  <div className="b3-extra">
-                    <div className="gs-slider-section">
-                      <div className="gs-slider-header">
-                        <span className="gs-slider-label">
-                          Beta Season Pool
-                        </span>
-                        <span className="gs-slider-value">
-                          {gsPercentage / 1000}%
-                        </span>
-                      </div>
-                      <input
-                        type="range"
-                        min={0}
-                        max={50000}
-                        step={1000}
-                        value={gsPercentage}
-                        onChange={(e) =>
-                          setGsPercentage(Number(e.target.value))
-                        }
-                        className="gs-slider-input"
-                        style={
-                          {
-                            "--gs-fill-pct": `${(gsPercentage / 50000) * 100}%`
-                          } as React.CSSProperties
-                        }
-                        disabled={isProcessing}
-                      />
-                      <div className="gs-slider-breakdown">
-                        <div className="gs-slider-breakdown-item">
-                          <span className="gs-slider-breakdown-label">
-                            Signal
-                          </span>
-                          <span className="gs-slider-breakdown-value">
-                            {formatTrust(breakdown.signalAmount)} TRUST
-                          </span>
-                        </div>
-                        <div className="gs-slider-breakdown-item">
-                          <span className="gs-slider-breakdown-label">
-                            Beta Season Pool
-                          </span>
-                          <span className="gs-slider-breakdown-value pool">
-                            {formatTrust(breakdown.poolAmount)} TRUST
-                          </span>
-                        </div>
-                      </div>
-                      {breakdown.belowMinimum && (
-                        <span className="gs-slider-minimum-hint">
-                          Below minimum — pool contribution skipped
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                )}
-
-                <div className="b3-ledger">
-                  <div className="b3-ledger-row">
-                    <span>Subtotal</span>
-                    <span>{formatTrust(breakdown.totalTrust)} TRUST</span>
-                  </div>
-                  {breakdown.poolAmount > 0 && (
-                    <div className="b3-ledger-row">
-                      <span>Beta Season Pool</span>
-                      <span>{formatTrust(breakdown.poolAmount)} TRUST</span>
-                    </div>
-                  )}
-                  {breakdown.platformPoolAmount > 0 && (
-                    <div className="b3-ledger-row">
-                      <span>{platformLabel}</span>
-                      <span>
-                        {formatTrust(breakdown.platformPoolAmount)} TRUST
-                      </span>
-                    </div>
-                  )}
-                  <div className="b3-ledger-row">
-                    <span>Fees</span>
-                    <span>{formatTrust(breakdown.totalFees)} TRUST</span>
-                  </div>
-                  <div className="b3-ledger-row b3-ledger-total">
-                    <span>Total deposit</span>
-                    <span>{formatTrust(breakdown.totalEstimate)} TRUST</span>
-                  </div>
-                  <div className="b3-ledger-row b3-ledger-row--muted">
-                    <span>Your balance</span>
-                    <span>{formatTrust(userBalance)} TRUST</span>
-                  </div>
-                </div>
-
-                <div className="b3-actions">
-                  <button
-                    className="b3-btn"
-                    type="button"
-                    onClick={handleClose}
-                    disabled={isProcessing}>
-                    Cancel
-                  </button>
-                  <button
-                    className="b3-btn b3-btn--primary"
-                    type="button"
-                    onClick={handleSubmit}
-                    disabled={
-                      isProcessing ||
-                      breakdown.totalEstimate > userBalance ||
-                      activeCount === 0
-                    }>
-                    {submitLabel || "Amplify"}
-                    <span className="b3-btn-amt">
-                      {formatTrust(breakdown.totalEstimate)} T
-                    </span>
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
+          <WeightAmplifyTicket
+            triplets={triplets}
+            removedIndices={removedIndices}
+            setRemovedIndices={setRemovedIndices}
+            onRemoveTriplet={onRemoveTriplet}
+            selectedWeights={selectedWeights}
+            activeCount={activeCount}
+            isProcessing={isProcessing}
+            processingStep={processingStep}
+            fixedDeposit={fixedDeposit}
+            ppEnabled={ppEnabled}
+            detectPlatformFromUrl={detectPlatformFromUrl}
+            getPpForSlug={getPpForSlug}
+            setPpForSlug={setPpForSlug}
+            onApplyAll={handleApplyAll}
+            onWeightSelect={handleWeightSelection}
+            curveSelector={curveSelector}
+            gsEnabled={gsEnabled}
+            gsPercentage={gsPercentage}
+            setGsPercentage={setGsPercentage}
+            breakdown={breakdown}
+            platformLabel={platformLabel}
+            userBalance={userBalance}
+            submitLabel={submitLabel}
+            onClose={handleClose}
+            onSubmit={handleSubmit}
+          />
         )}
 
         {rewardClaimed && discoveryReward && (
-          <div className="rc">
-            <div className="rc-ember" />
-            <div className="rc-ticket">
-              <div className="rc-stub">
-                <div className="rc-stub-row">
-                  <span className="rc-stub-k">N°</span>
-                  <span className="rc-stub-v">
-                    {transactionHash
-                      ? transactionHash.slice(2, 10).toUpperCase()
-                      : "—"}
-                  </span>
-                </div>
-                <div className="rc-stub-row">
-                  <span className="rc-stub-k">DATE</span>
-                  <span className="rc-stub-v">
-                    {new Date()
-                      .toLocaleDateString("fr-FR")
-                      .replace(/\//g, " · ")}
-                  </span>
-                </div>
-              </div>
-              <div className="rc-perf" />
-
-              <div className="rc-body">
-                <div className="rc-headline">
-                  <span className="rc-drop">M</span>
-                  <h1 className="rc-h1">
-                    ark
-                    <br />
-                    captured.
-                  </h1>
-                </div>
-                <p className="rc-sub">
-                  {keptTriplets.length === 1
-                    ? "Awarded for marking one page."
-                    : `Awarded for marking ${keptTriplets.length} pages.`}
-                </p>
-
-                <div className="rc-marks">
-                  {keptTriplets.map((t) => (
-                    <div className="rc-mark" key={t.id}>
-                      <span className="rc-mark-verb">
-                        {t.triplet.predicate}
-                      </span>
-                      <span className="rc-mark-obj" title={t.triplet.object}>
-                        {t.description || t.triplet.object}
-                      </span>
-                      <span className="rc-mark-meta">
-                        {(t.interestContexts?.length
-                          ? t.interestContexts
-                          : t.interestContext
-                            ? [t.interestContext]
-                            : []
-                        ).map((slug) => {
-                          const l = contextLabel(slug)
-                          return l ? (
-                            <span className="rc-mark-ctx" key={slug}>
-                              {l}
-                            </span>
-                          ) : null
-                        })}
-                        <span className="rc-mark-host">
-                          {hostFromUrl(t.url)}
-                        </span>
-                      </span>
-                    </div>
-                  ))}
-                </div>
-
-                <div className="rc-tiers">
-                  <div className="rc-tiers-k">Tier ladder</div>
-                  {TIER_ORDER.map((tier) => {
-                    const earned = discoveryReward.status.toLowerCase() === tier
-                    return (
-                      <div
-                        className={`rc-tier${earned ? " is-earned" : ""}`}
-                        key={tier}>
-                        <UserBadge
-                          tier={tier as RewardTier}
-                          iconUrl={BADGE_IMAGES[tier as RewardTier]}
-                          size={32}
-                        />
-                        <span className="rc-tier-label">
-                          {tier.charAt(0).toUpperCase() + tier.slice(1)}
-                        </span>
-                        <span className="rc-tier-state">
-                          {earned ? "earned" : "locked"}
-                        </span>
-                      </div>
-                    )
-                  })}
-                </div>
-
-                <div className="rc-ledger">
-                  <div className="rc-ledger-row">
-                    <span>Gold reward</span>
-                    <span>+ {discoveryReward.gold} G</span>
-                  </div>
-                  <div className="rc-ledger-row rc-ledger-total">
-                    <span>Balance</span>
-                    <span>{totalGold.toLocaleString()} G</span>
-                  </div>
-                </div>
-
-                <div className="rc-actions">
-                  {transactionHash ? (
-                    <a
-                      className="rc-btn rc-btn--primary"
-                      href={`${EXPLORER_URLS.TRANSACTION}${transactionHash}`}
-                      target="_blank"
-                      rel="noopener noreferrer">
-                      View my Echo →
-                    </a>
-                  ) : (
-                    <button
-                      className="rc-btn rc-btn--primary"
-                      onClick={handleClose}>
-                      View my Echo →
-                    </button>
-                  )}
-                  <a
-                    className="rc-btn"
-                    href={`https://x.com/intent/tweet?text=${encodeURIComponent(
-                      `Just marked ${keptTriplets.length} page${
-                        keptTriplets.length === 1 ? "" : "s"
-                      } on Sofia and earned ${discoveryReward.gold} Gold.`
-                    )}`}
-                    target="_blank"
-                    rel="noopener noreferrer">
-                    Share on X
-                  </a>
-                  <button className="rc-btn" onClick={handleClose}>
-                    Done
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
+          <WeightRewardTicket
+            discoveryReward={discoveryReward}
+            transactionHash={transactionHash}
+            keptTriplets={keptTriplets}
+            totalGold={totalGold}
+            onClose={handleClose}
+          />
         )}
       </div>
     </div>,
