+++ packages/design-system/src/lib/index.ts
@@ -1,1 +1,2 @@
+export * from './chipOverflow'
 export * from './formatDuration'
